Fix encoding of short vs integer
* encode/decode `short` values with AsnUint8Codec
* encode/decode `int` values with AsnUint16Codec
* Fix unit tests.
* Improve Javadoc and formatting.
* Remove duplicate code.

+++ codec-framework/src/main/java/org/interledger/encoding/asn/serializers/oer/AsnUint8OerSerializer.java
@@ -9,9 +9,9 @@ package org.interledger.encoding.asn.serializers.oer;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -31,8 +31,7 @@ import java.io.OutputStream;
 import java.util.Objects;
 
 /**
- * An extension of {@link AsnObjectSerializer} for reading and writing an ASN.1 OER 8-Bit
- * integer.
+ * An extension of {@link AsnObjectSerializer} for reading and writing an ASN.1 OER 8-Bit integer.
  *
  * <p>This serializer is used instead of the generic Octet String serializer for performance
  * reasons.
@@ -41,14 +40,13 @@ public class AsnUint8OerSerializer implements AsnObjectSerializer<AsnUint8Codec>
 
   @Override
   public void read(final AsnObjectSerializationContext context, final AsnUint8Codec instance,
-                   final InputStream
-      inputStream)
-      throws IOException {
+      final InputStream inputStream
+  ) throws IOException {
     Objects.requireNonNull(context);
     Objects.requireNonNull(instance);
     Objects.requireNonNull(inputStream);
 
-    int value = inputStream.read();
+    short value = (short) inputStream.read();
     if (value == -1) {
       throw new CodecException("Unexpected end of stream.");
     }
@@ -59,7 +57,7 @@ public class AsnUint8OerSerializer implements AsnObjectSerializer<AsnUint8Codec>
 
   @Override
   public void write(AsnObjectSerializationContext context, AsnUint8Codec instance,
-                    OutputStream outputStream)
+      OutputStream outputStream)
       throws IOException {
 
     if (instance.decode() > 255) {
